Commands

The main command is xps, by default it installs to JULIA_DEPOT_PATH/bin.

The first-order commands are

xps
├── print
└── run

The usage are

ExpressCommands.print — Function

Print the file in a pretty format.

Arguments

file: the file to be printed. Available extensions are .jls, .jld2, .json, .yaml, .yml or .toml.

Missing docstring.

Missing docstring for run. Check Documenter's build log for details.

And the secondary commands are

xps eos
    ├── fit
    └── plot

ExpressCommands.EOS.fit — Function

Fit an equation of state from cfgfile for calculation calc.

Arguments

calc: the calculation type. Acceptable options are "scf" for self-consistent calculation and "optim" for structure optimizations.

cfgfile: the file to be printed. Available extensions are .jls, .json, .yaml, .yml or .toml.

ExpressCommands.EOS.plot — Function

Plot an equation of state from file to file "out".

Arguments

file: a .jls file that saves the equation of state.