Commands

The main command is xps, by default it installs to JULIA_DEPOT_PATH/bin.

The first-order commands are

xps
├── print
└── run
└── graph

The usage are

ExpressCommands.print — Function

Print the file in a pretty format.

Args

file: the file to be printed. Available extensions are .jld2, .json, .yaml, .yml or .toml.

ExpressCommands.run — Function

Run a configuration file (with an absolute path will be better).

Args

file: the file to be run. Acceptable extensions are .json, .yaml, .yml, or .toml.

Missing docstring.

Missing docstring for graph. Check Documenter's build log for details.

You can always type

xps --help

to see the help information, or type

xps --version

to see the current version of xps.

The equation of state workflow commands

The secondary commands are

xps eos
    ├── run
    └── plot
    └── fit

ExpressCommands.EOS.run — Function

Run a configuration file (with an absolute path will be better). It is equivalent to xps run <file>.

Args

file: the file to be run. Acceptable extensions are .json, .yaml, .yml, or .toml.

ExpressCommands.EOS.fit — Function

Fit an equation of state from a configuration file for calculation calc.

Args

calc: the calculation type. Acceptable options are "scf" for self-consistent calculation and "optim" for structure optimizations.

file: the file which saves the raw data. Available extensions are .jld2, .json, .yaml, .yml or .toml.

Missing docstring.

Missing docstring for ExpressCommands.EOS.plot. Check Documenter's build log for details.

The phonon workflow commands

The secondary commands are

xps ph
    └── run

ExpressCommands.Ph.run — Function

Run a configuration file (with an absolute path will be better). It is equivalent to xps run <file>.

Args

file: the file to be run. Acceptable extensions are .json, .yaml, .yml, or .toml.

The QHA workflow commands

The secondary commands are

xps qha
    └── run

ExpressCommands.QHA.run — Function

Run a configuration file (with an absolute path will be better). It is equivalent to xps run <file>.

Args

file: the file to be run. Acceptable extensions are .json, .yaml, .yml, or .toml.